modified functionality of pick/picked activity apply

+++ icarving.api.pinche/src/main/java/cn/icarving/api/pinche/dao/BaseDao.java
@@ -55,7 +55,7 @@ public class BaseDao<T> {
 	}
 
 	@SuppressWarnings("unchecked")
-	public T find(int id) {
+	public T find(long id) {
 		Session session = sessionFactory.openSession();
 		session.getTransaction().begin();
 		T result = (T) session.get(type, id);
